backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
neo4j==5.25.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.35.0
python-dotenv==1.0.1
alembic==1.14.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2